:recycle: Rename `Deposit` model to `SystemTransaction`
